pyproject.toml
[project]
name = "signal-trck"
version = "0.1.0"
description = "Personal crypto charting + LLM-grounded technical analysis"
authors = [{ name = "Hugues Lagrange" }]
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "aiosqlite>=0.20",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "structlog>=24.0",
    "typer>=0.12",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-recording>=0.13",
    "ruff>=0.8",
]

[project.scripts]
signal-trck = "signal_trck.cli.main:app"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
